Split live providers into chunks.
