+++ .github/workflows/release.yml
@@ -4,6 +4,7 @@ name: Release
 on:
   release:
     types: [published]
+  workflow_dispatch:
 
 jobs:
   release:
@@ -18,14 +19,14 @@ jobs:
           yarn install
           yarn run build
 
-      # Upload build file to the releas as an asset.
+      # Upload build file to the release as an asset.
       - name: Upload zip to release
         uses: svenstaro/upload-release-action@v1-release
 
         with:
           repo_token: ${{ secrets.GITHUB_TOKEN }}
-          file: dist/frigate-hass-card.js
-          asset_name: frigate-hass-card.js
+          file: dist/*.js
+          file_glob: true
           tag: ${{ github.ref }}
           overwrite: true
