+++ src/card.ts
@@ -27,8 +27,8 @@ import { FrigateCardElements } from './components/elements.js';
 import './components/gallery.js';
 import './components/image.js';
 import { FrigateCardImage } from './components/image.js';
-import './components/live.js';
-import { FrigateCardLive } from './components/live.js';
+import './components/live/live.js';
+import type { FrigateCardLive } from './components/live/live.js';
 import './components/menu.js';
 import { FrigateCardMenu, FRIGATE_BUTTON_MENU_ICON } from './components/menu.js';
 import './components/message.js';
